+++ android/app/src/main/java/com/woowacourse/ody/data/remote/repository/DefaultNotificationLogRepository.kt
@@ -1,6 +1,8 @@
-package com.woowacourse.ody.data.remote
+package com.woowacourse.ody.data.remote.repository
 
 import com.woowacourse.ody.data.model.toNotificationList
+import com.woowacourse.ody.data.remote.RetrofitClient
+import com.woowacourse.ody.data.remote.service.NotificationService
 import com.woowacourse.ody.domain.NotificationLog
 import com.woowacourse.ody.domain.NotificationLogRepository
 
@@ -8,5 +10,6 @@ object DefaultNotificationLogRepository : NotificationLogRepository {
     private val notificationService: NotificationService =
         RetrofitClient.retrofit.create(NotificationService::class.java)
 
-    override suspend fun getNotificationLogs(): List<NotificationLog> = notificationService.getNotificationLogs().toNotificationList()
+    override suspend fun fetchNotificationLogs(meetingId: Int): List<NotificationLog> =
+        notificationService.fetchNotificationLogs(meetingId).toNotificationList()
 }
