+++ android/app/src/main/java/com/woowacourse/ody/data/remote/service/NotificationService.kt
@@ -1,11 +1,14 @@
-package com.woowacourse.ody.data.remote
+package com.woowacourse.ody.data.remote.service
 
 import com.woowacourse.ody.data.model.notification.NotificationLogEntities
 import retrofit2.http.GET
+import retrofit2.http.Path
 
 interface NotificationService {
     @GET(NOTIFICATION_LOG_PATH)
-    suspend fun getNotificationLogs(): NotificationLogEntities
+    suspend fun fetchNotificationLogs(
+        @Path("meetingId") meetingId: Int,
+    ): NotificationLogEntities
 
     companion object {
         const val NOTIFICATION_LOG_PATH = "/meetings/{meetingId}/noti-log"
