+++ android/app/src/main/java/com/woowacourse/ody/domain/NotificationLogRepository.kt
@@ -1,5 +1,5 @@
 package com.woowacourse.ody.domain
 
 interface NotificationLogRepository {
-    suspend fun getNotificationLogs(): List<NotificationLog>
+    suspend fun fetchNotificationLogs(meetingId: Int): List<NotificationLog>
 }
