+++ android/app/src/main/java/com/woowacourse/ody/presentation/notificationlog/NotificationLogViewModel.kt
@@ -4,26 +4,41 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.woowacourse.ody.domain.MeetingRepository
 import com.woowacourse.ody.domain.NotificationLogRepository
+import com.woowacourse.ody.presentation.notificationlog.uimodel.MeetingUiModel
 import com.woowacourse.ody.presentation.notificationlog.uimodel.NotificationLogUiModel
+import com.woowacourse.ody.presentation.notificationlog.uimodel.toMeetingUiModel
 import com.woowacourse.ody.presentation.notificationlog.uimodel.toNotificationUiModels
 import kotlinx.coroutines.launch
 
 class NotificationLogViewModel(
     private val notificationLogRepository: NotificationLogRepository,
+    private val meetingRepository: MeetingRepository,
 ) : ViewModel(), CopyInviteCodeButtonListener {
+    private val _meeting = MutableLiveData<MeetingUiModel>()
+    val meeting: LiveData<MeetingUiModel> = _meeting
+
     private val _notificationLogs = MutableLiveData<List<NotificationLogUiModel>>()
     val notificationLogs: LiveData<List<NotificationLogUiModel>> = _notificationLogs
 
-    private fun fetchNotificationLogs() =
+    private fun fetchNotificationLogs(meetingId: Int) =
         viewModelScope.launch {
-            notificationLogRepository.getNotificationLogs().let { notificationLogs ->
+            notificationLogRepository.fetchNotificationLogs(meetingId).let { notificationLogs ->
                 _notificationLogs.postValue(notificationLogs.toNotificationUiModels())
             }
         }
 
+    private fun fetchMeeting() =
+        viewModelScope.launch {
+            meetingRepository.fetchMeeting().let { meeting ->
+                _meeting.postValue(meeting.toMeetingUiModel())
+            }
+        }
+
     fun initialize() {
-        fetchNotificationLogs()
+        fetchNotificationLogs(0)
+        fetchMeeting()
     }
 
     override fun onClickCopyInviteCode() {
